requirements.txt
torch>=2.0.0
transformers==4.44.2
peft>=0.4.0
accelerate>=0.20.0
datasets>=2.14.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.65.0
einops>=0.6.0
bitsandbytes>=0.41.0
scipy>=1.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0 